move avg_embeddings to nlp_tools
